pr => moved queries to resources queries file
